requirements.txt
# Core dependencies
requests>=2.31.0

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# Git operations
gitpython>=3.1.40

# CLI and output
rich>=13.5.0
click>=8.1.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Code analysis
astroid>=2.15.0
radon>=6.0.1

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0